requirements.txt
azure-core
azure-ai-formrecognizer
azure-ai-documentintelligence
python-dotenv
openai
colorama